twitter-sentiment-analysis/requirements.txt
Flask==2.2.2
pandas==1.4.1
numpy==1.22.3
spacy==3.2.4
vadersentiment==3.3.2
unicodedata2==14.0.0
PyYAML==6.0
contractions==0.1.68
tweepy== 4.8.0
nltk==3.7
pyodbc==4.0.34


